+++ packages/app-builder-lib/src/node-module-collector/npmNodeModulesCollector.ts
@@ -1,3 +1,4 @@
+import { Lazy } from "lazy-val"
 import { NodeModulesCollector } from "./nodeModulesCollector"
 import { DependencyTree, NpmDependency, ParsedDependencyTree } from "./types"
 import { log } from "builder-util"
@@ -7,11 +8,10 @@ export class NpmNodeModulesCollector extends NodeModulesCollector<NpmDependency,
     super(rootDir)
   }
 
-  getCommand(): string {
-    return process.platform === "win32" ? "npm.cmd" : "npm"
-  }
+  public readonly pmCommand = new Lazy<string>(() => Promise.resolve(process.platform === "win32" ? "npm.cmd" : "npm"))
+  public readonly installOptions = this.pmCommand.value.then(cmd => ({ cmd, args: ["ci"], lockfile: "package-lock.json" }))
 
-  getArgs(): string[] {
+  protected getArgs(): string[] {
     return ["list", "-a", "--include", "prod", "--include", "optional", "--omit", "dev", "--json", "--long", "--silent"]
   }